Brings the two latest fixes from claude/quirky-mendel-my4tvl onto main:
- tolerate string-encoded Monitor G5 entity IDs (parsing crash fix)
- auto-relogin on expired session (401) with one retry

+++ internal/sickan/monitor.go
@@ -168,7 +168,7 @@ func (tb *Toolbox) monitorFillMissingCertData(input json.RawMessage) (string, er
 		if i >= 10 {
 			break
 		}
-		m := match{PartId: r.PartId, SerialNumber: r.SerialNumber}
+		m := match{PartId: int64(r.PartId), SerialNumber: r.SerialNumber}
 		if r.PurchaseOrderId != 0 {
 			if po, _ := tb.Monitor.GetPurchaseOrder(ctx, r.PurchaseOrderId); po != nil {
 				m.OrderNumber = po.OrderNumber
